feat: curve adapter adjustment + price feed for curve crypto

+++ contracts/test/adapters/mainnet/Live_SUSDEquivalenceTest.sol
@@ -93,7 +93,7 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
             CurveV1Multicaller deposit = CurveV1Multicaller(curveDepositAddr);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(pool.exchange(0, 1, 2000 * WAD, 1500 * (10 ** 6))));
+            creditFacade.multicall(multicallBuilder(pool.exchange(int128(0), int128(1), 2000 * WAD, 1500 * (10 ** 6))));
             comparator.takeSnapshot("after_exchange", accountToSaveBalances);
 
             uint256[4] memory amounts = [1000 * WAD, 1000 * (10 ** 6), 0, 0];
@@ -109,7 +109,7 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
             comparator.takeSnapshot("after_remove_liquidity", accountToSaveBalances);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(deposit.remove_liquidity_one_coin(500 * WAD, 1, 0)));
+            creditFacade.multicall(multicallBuilder(deposit.remove_liquidity_one_coin(500 * WAD, int128(1), 0)));
             comparator.takeSnapshot("after_remove_liquidity_one_coin", accountToSaveBalances);
 
             amounts = [500 * WAD, 100 * (10 ** 6), 0, 0];
@@ -119,26 +119,26 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
             comparator.takeSnapshot("after_remove_liquidity_imbalance", accountToSaveBalances);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(pool.add_liquidity_one_coin(100 * WAD, 0, 50 * WAD)));
+            creditFacade.multicall(multicallBuilder(pool.add_liquidity_one_coin(100 * WAD, int128(0), 50 * WAD)));
             comparator.takeSnapshot("after_add_liquidity_one_coin", accountToSaveBalances);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(pool.exchange_all(0, 1, RAY / 2 / 10 ** 12)));
+            creditFacade.multicall(multicallBuilder(pool.exchange_all(int128(0), int128(1), RAY / 2 / 10 ** 12)));
             comparator.takeSnapshot("after_exchange_all", accountToSaveBalances);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(pool.add_all_liquidity_one_coin(1, (RAY * 10 ** 12) / 2)));
+            creditFacade.multicall(multicallBuilder(pool.add_all_liquidity_one_coin(int128(1), (RAY * 10 ** 12) / 2)));
             comparator.takeSnapshot("after_add_all_liquidity_one_coin", accountToSaveBalances);
 
             evm.prank(USER);
-            creditFacade.multicall(multicallBuilder(deposit.remove_all_liquidity_one_coin(0, RAY / 2)));
+            creditFacade.multicall(multicallBuilder(deposit.remove_all_liquidity_one_coin(int128(0), RAY / 2)));
             comparator.takeSnapshot("after_remove_all_liquidity_one_coin", accountToSaveBalances);
         } else {
             ICurvePool4Assets pool = ICurvePool4Assets(curvePoolAddr);
             ICurvePool4Assets deposit = ICurvePool4Assets(curveDepositAddr);
 
             evm.prank(USER);
-            pool.exchange(0, 1, 2000 * WAD, 1500 * (10 ** 6));
+            pool.exchange(int128(0), int128(1), 2000 * WAD, 1500 * (10 ** 6));
             comparator.takeSnapshot("after_exchange", accountToSaveBalances);
 
             uint256[4] memory amounts = [1000 * WAD, 1000 * (10 ** 6), 0, 0];
@@ -154,7 +154,7 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
             comparator.takeSnapshot("after_remove_liquidity", accountToSaveBalances);
 
             evm.prank(USER);
-            deposit.remove_liquidity_one_coin(500 * WAD, 1, 0);
+            deposit.remove_liquidity_one_coin(500 * WAD, int128(1), 0);
             comparator.takeSnapshot("after_remove_liquidity_one_coin", accountToSaveBalances);
 
             amounts = [500 * WAD, 100 * (10 ** 6), 0, 0];
@@ -169,7 +169,7 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
 
             uint256 balanceToSwap = tokenTestSuite.balanceOf(Tokens.DAI, accountToSaveBalances) - 1;
             evm.prank(USER);
-            pool.exchange(0, 1, balanceToSwap, balanceToSwap / (2 * 10 ** 12));
+            pool.exchange(int128(0), int128(1), balanceToSwap, balanceToSwap / (2 * 10 ** 12));
             comparator.takeSnapshot("after_exchange_all", accountToSaveBalances);
 
             balanceToSwap = tokenTestSuite.balanceOf(Tokens.USDC, accountToSaveBalances) - 1;
@@ -179,7 +179,7 @@ contract Live_CurveSusdEquivalenceTest is DSTest, LiveEnvHelper {
 
             balanceToSwap = tokenTestSuite.balanceOf(Tokens.crvPlain3andSUSD, accountToSaveBalances) - 1;
             evm.prank(USER);
-            deposit.remove_liquidity_one_coin(balanceToSwap, 0, balanceToSwap / 2);
+            deposit.remove_liquidity_one_coin(balanceToSwap, int128(0), balanceToSwap / 2);
             comparator.takeSnapshot("after_remove_all_liquidity_one_coin", accountToSaveBalances);
         }
     }
